+++ lib/luna_park/errors.rb
@@ -5,5 +5,6 @@ module LunaPark
     class Processing     < StandardError; end
     class AbstractMethod < RuntimeError; end
     class Unwrapable     < ArgumentError; end
+    class NotConfigured  < StandardError; end
   end
 end