+++ lib/luna_park/use_cases/command.rb
@@ -19,6 +19,8 @@ module LunaPark
       private
 
       # :nocov:
+
+      # @abstract
       def execute
         raise Errors::AbstractMethod
       end