+++ lib/luna_park/values/compound.rb
@@ -3,27 +3,20 @@
 module LunaPark
   module Values
     class Compound
+      extend  Extensions::Wrappable
       include Extensions::Attributable
 
-      def initialize(attrs)
+      def initialize(attrs = {})
         set_attributes attrs
       end
 
       # :nocov:
+
+      # @abstract
       def ==(_other)
         raise Errors::AbstractMethod
       end
       # :nocov:
-
-      class << self
-        def wrap(input)
-          case input
-          when self then input
-          when Hash then new(input)
-          else raise Errors::Unwrapable, "#{self} can`t wrap #{input.class}"
-          end
-        end
-      end
     end
   end
 end